fetches discovery document to get endpoints
I think i solved how to get the keys but I do not think fething the
discorevydocument like this is good. The http-response from the
discovery endpoint will not be cached by the browser so it is clearly
not meant to be used like this. The key response though has a
cache-control. I will refactor to use a hardcoded endpoints through
flags. The keys will be fetched though.