Adds steps to login-flow
if login-redirect
1. parse the state and send it to js. The value of the state parameter
is a key to sessionstorage. Right now it stores the nonce but it could
store more data, say a route we wanted to follow but that needed
login-status.
2. If we get some state value back we use it to validate the nonce in
the token.
3. We parse the kid from the token and calls the key-endpoint to obtain
the keyset from the auth-server. This response is normally cached for as
long as the key is valid but the advice is to take into account that
keys might need to be rolled over at any time.
4. If on the first try we cannot find the key in the keyset we make the
request one more time this time with no-cache header to make sure we get
the latest keyset.
5. We send the token together with its key to js where it is validated
using jsrsasign lib
6. When we get the token back we are officially logged in.