maps a current host to model
Makes it a little bit nicer to use the interface if we have a
representation in the view of the current host. The hostId is an
argument to the appointments api.

I also started thinking about interacting with the api and what is the
best way to think about time and dates.

as for the api I'm thinking
GET host/HostId/appoinments?from=unixTimestamp&dur=ms
OK 200
{[{ start: unixTimestamp, dur: float},
               ...
]}

I made a little appempt to get unixTimestamp of the start of the current
week. But it would be most interesting if I get an actual api up and
running serving mockdata.